Move caseEquals() to tests, fix test for it
